Restructured structure of related classes
