+++ src/main/java/cybuy/item/ItemRepository.java
@@ -1,4 +1,4 @@
-package cybuy.data;
+package cybuy.item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
